Refactor getImportPath
This consolidates the functionality found in two commands into one place
so that if it ever needs to be changed or removed it can be done more
easily.
