+++ src/goGetImport.ts
@@ -5,7 +5,7 @@ import cp = require('child_process');
 import path = require('path');
 import fs = require('fs');
 import { getGoRuntimePath } from './goPath';
-import { getCurrentGoPath } from './util';
+import { getCurrentGoPath, getImportPath } from './util';
 
 export function goGetImport() {
 	const editor = vscode.window.activeTextEditor;
@@ -35,20 +35,3 @@ export function goGetImport() {
 		});
 	});
 };
-
-// TODO: maybe move this to some util package? It's copied from goBrowsePackages.ts
-function getImportPath(text: string): string {
-	// Catch cases like `import alias "importpath"` and `import "importpath"`
-	let singleLineImportMatches = text.match(/^\s*import\s+([a-z,A-Z,_,\.]\w*\s+)?\"([^\"]+)\"/);
-	if (singleLineImportMatches) {
-		return singleLineImportMatches[2];
-	}
-
-	// Catch cases like `alias "importpath"` and "importpath"
-	let groupImportMatches = text.match(/^\s*([a-z,A-Z,_,\.]\w*\s+)?\"([^\"]+)\"/);
-	if (groupImportMatches) {
-		return groupImportMatches[2];
-	}
-
-	return text.trim();
-}
